+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 1.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Drop support for Python 3.6 & 3.7
 
 
 1.1.0 (2021-06-23)